renamed elarateam.products package
