--- Project-Elara/elara-admin-console/src/main/java/ru/hse/elarateam/ElaraAdminConsoleApplication.java
@@ -1,10 +1,0 @@
-package ru.hse.elarateam;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ElaraAdminConsoleApplication {
-    public static void main(String[] args) {
-        org.springframework.boot.SpringApplication.run(ElaraAdminConsoleApplication.class, args);
-    }
-}
