+++ Project-Elara/elara-products/src/main/java/ru/hse/products/model/Color.java
@@ -1,4 +1,4 @@
-package ru.hse.elarateam.model;
+package ru.hse.products.model;
 
 import jakarta.persistence.*;
 import lombok.*;
